move batch iterator
